complete github authentication
