+++ src/app/page.tsx
@@ -1,6 +1,6 @@
 import { createServerComponentClient } from "@supabase/auth-helpers-nextjs";
 import { cookies } from "next/headers";
-import { AuthButton } from "@/components/AuthButton";
+import GithubAuthButtonServer from "@/components/auth/github/GithubAuthButtonServer";
 
 export default async function Home() {
   const supabase = createServerComponentClient({ cookies });
@@ -8,7 +8,7 @@ export default async function Home() {
 
   return (
     <>
-      <AuthButton />
+      <GithubAuthButtonServer />
       <pre>{JSON.stringify(posts, null, 2)}</pre>;
     </>
   );
